Turn shoulddointerception into a recipe function
